Don't copy the source files if they haven't changed
If the source files in the OpenBSD version have not changed since the
last time update.sh was run, the script will still override them,
updating the mtime and making the build system think it needs to rebuild
the whole library even if a single file changed.

Change the copy command to compare the files and only overwrite the
target if it has changed.