Use Critical Sections Instead
First of all, i see crypto_init_locks is not being called anywhere in the entire library, the change above will not work until that is fixed, it will just crash because of the uninitialized memory.
I'm pretty sure the original code is not doing anything either, since the HANDLE array is not set, WaitForSingleObject is firing instantly without waiting.

Apart from that issue, i would recommend the change above since Critical Sections are less costly and faster. They are faster because Mutex always goes directly for a kernel lock, while Critical Sections will do the equivalent of InterlockedIncrement, and if it succeeds then it considers the lock acquired without having gone into the kernel. It is also possible to add some spin to the Critical Sections, but it is detrimental on Hyper-Threading systems according to this: https://software.intel.com/en-us/articles/managing-lock-contention-large-and-small-critical-sections so i did not include it. OpenSSL does include the spin count in their code.

Mutexes should also only be used when it is required that the lock be accessible over multiple processes, but looking at the linux code of this, it is not the case (since in that code the pthreads mutex was not configured for shared access). On top of all that, WaitForSingleObject has a possibility of deadlocking under certain conditions.

I could also not find any reference to functions releasing the Mutexes, so that memory will leak on program close as well (on linux too, i didn't find any pthread_mutex_destroy). I included a free function for the Critical Sections at least that should be called somewhere in the library on shutdown.