tests: Fix the build with --disable-static
The tests require --enable-static because -export-symbols is set in
libtls_la_LDFLAGS which hides many of the internal symbols for the shared
library.

This can be avoided by compiling most of the code into a convenience
library which the tests and libtls can both link against. The installed
shared library will still use -export-symbols as expected without
conflicting with the tests that require access to internal symbols.

Fixes https://github.com/libressl-portable/portable/issues/754