cmake: allow cross-CPU builds via `LIBRESSL_CPU`
If `LIBRESSL_CPU` is not set, honor existing `CMAKE_OSX_ARCHITECTURE`
for Apple targets. Then fall back to `CMAKE_C_COMPILER_TARGET` if set.
Otherwise fall back to the value that was used before this patch:
`CMAKE_SYSTEM_PROCESSOR`.

This allows customizing the target CPU, e.g. on Linux.

I'm not certain of the exact CMake rules for this, but on some platforms
(e.g. Linux) it is not possible to override the default
`CMAKE_SYSTEM_PROCESSOR` value by passing it to cmake via
`-DCMAKE_SYSTEM_PROCESSOR=...`. Meaning, before this patch it was not
possible to build for a CPU other than the host one.

Also:
- simplify syntax in `STREQUAL` and `MATCH` expressions in the lines
  touched.
- display the target CPU and ABI in the configuration phase.