Windows: Fix assertion pop-up up when using Debug compiled libressl.
When running the signertest, or the test project in 
https://github.com/libressl/portable/issues/266 an assertion window
pops up. This was fixed in afcd4be8a72a for a release compiled library.
To prevent the issue in debug mode, it looks like it is necessary to 
also disable the assertion window popup. 

With this all tests pass when compiling and running them with a Debug,
Release or RelWithDebInfo CMake build on windows (for me).