Remove the dist-win compilation script
With the CMake version of building for windows
there is support for building with MINGW, clang,
and Visual Studio. CMake can build for MINGW
using the regular method of autogen and CMake.

Removing this makes it a bit simpler to get an
overview of how to build for windows. Reducing
confusion.