Try to simplify build instruction.
In particular simplify the Windows build instructions,
by only having a short CMake section about it. While
mentioning that bash shell is required also on Windows
up-front.

Splits the build section into a section for configure
and one for CMake so it is easier to pick one of them.
Moves the additional CMake options into the CMake section.
Moves dist.sh text into the configure section as it is
configure specific.