Clean up the time_t test mess
The RFC 5280 test now passes also with small time_t. The ASN.1 time
test now has a test case that makes it fail for small time t. In that
case use a wrapper script that prins why the test is expected to fail
and makes the test suite fail if it passes.