fix gitignore brokenness in crypto
/crypto/* masks too much and breaks the exclusion function.

The docs for gitignore state that if a parent directory is ignored,
negating a pattern with a ! prefix has no effect since it is not
possible to reinclude a file if the parent director is excluded.

https://git-scm.com/docs/gitignore

This just is more explicit, which catches files added out of tree more
readily.

Also delete the long-removed libtls-standalone references